fix: servertimeoffset

+++ src/main/java/org/torusresearch/torusutils/apis/KeyLookupResult.java
@@ -13,14 +13,14 @@ public class KeyLookupResult {
     public final KeyResult keyResult;
 
     public final List<Integer> nodeIndexes;
-    public final BigInteger server_time_offset;
+    public final Integer server_time_offset;
     @Nullable
     public final GetOrSetNonceResult nonceResult;
 
     @Nullable
     public final JRPCResponse.ErrorInfo errorResult;
 
-    public KeyLookupResult(@Nullable KeyResult keyResult, List<Integer> nodeIndexes, BigInteger server_time_offset, @Nullable GetOrSetNonceResult nonceResult, @Nullable JRPCResponse.ErrorInfo errorResult) {
+    public KeyLookupResult(@Nullable KeyResult keyResult, List<Integer> nodeIndexes, Integer server_time_offset, @Nullable GetOrSetNonceResult nonceResult, @Nullable JRPCResponse.ErrorInfo errorResult) {
         this.keyResult = keyResult;
         this.nodeIndexes = nodeIndexes;
         this.server_time_offset = server_time_offset;
